consolidate utils: remove dev/prod split, centralize contract files
- Remove src/utils/dev/ and src/utils/prod/ directories
- Move contract initialization to src/utils/contractInit.ts
- Move wallet initialization to src/utils/walletInit.ts
- Move error handling to src/utils/handleContractError.ts
- Simplifies import paths and removes environment-based duplication
