--- packages/dapp/src/utils/prod/abis/EventEscrow.json
@@ -1,163 +1,0 @@
-[
-   {
-      "inputs": [
-         {
-            "internalType": "contract IERC20",
-            "name": "_usdtToken",
-            "type": "address"
-         },
-         {
-            "internalType": "address[]",
-            "name": "_beneficiaries",
-            "type": "address[]"
-         },
-         {
-            "internalType": "uint256[]",
-            "name": "_shares",
-            "type": "uint256[]"
-         }
-      ],
-      "stateMutability": "nonpayable",
-      "type": "constructor"
-   },
-   {
-      "anonymous": false,
-      "inputs": [
-         {
-            "indexed": false,
-            "internalType": "address",
-            "name": "usdtToken",
-            "type": "address"
-         },
-         {
-            "indexed": false,
-            "internalType": "address[]",
-            "name": "beneficiaries",
-            "type": "address[]"
-         },
-         {
-            "indexed": false,
-            "internalType": "uint256[]",
-            "name": "shares",
-            "type": "uint256[]"
-         },
-         {
-            "indexed": false,
-            "internalType": "uint256",
-            "name": "numBeneficiaries",
-            "type": "uint256"
-         }
-      ],
-      "name": "EventEscrowDeployed",
-      "type": "event"
-   },
-   {
-      "anonymous": false,
-      "inputs": [
-         {
-            "indexed": false,
-            "internalType": "address",
-            "name": "beneficiary",
-            "type": "address"
-         },
-         {
-            "indexed": false,
-            "internalType": "uint256",
-            "name": "amount",
-            "type": "uint256"
-         },
-         {
-            "indexed": false,
-            "internalType": "address",
-            "name": "indexedcaller",
-            "type": "address"
-         }
-      ],
-      "name": "PaymentDistributed",
-      "type": "event"
-   },
-   {
-      "inputs": [
-         {
-            "internalType": "uint256",
-            "name": "",
-            "type": "uint256"
-         }
-      ],
-      "name": "beneficiaries",
-      "outputs": [
-         {
-            "internalType": "address",
-            "name": "",
-            "type": "address"
-         }
-      ],
-      "stateMutability": "view",
-      "type": "function"
-   },
-   {
-      "inputs": [],
-      "name": "distributePayments",
-      "outputs": [],
-      "stateMutability": "nonpayable",
-      "type": "function"
-   },
-   {
-      "inputs": [],
-      "name": "numBeneficiaries",
-      "outputs": [
-         {
-            "internalType": "uint256",
-            "name": "",
-            "type": "uint256"
-         }
-      ],
-      "stateMutability": "view",
-      "type": "function"
-   },
-   {
-      "inputs": [
-         {
-            "internalType": "address",
-            "name": "",
-            "type": "address"
-         }
-      ],
-      "name": "payouts",
-      "outputs": [
-         {
-            "internalType": "uint256",
-            "name": "",
-            "type": "uint256"
-         }
-      ],
-      "stateMutability": "view",
-      "type": "function"
-   },
-   {
-      "inputs": [],
-      "name": "totalShares",
-      "outputs": [
-         {
-            "internalType": "uint256",
-            "name": "",
-            "type": "uint256"
-         }
-      ],
-      "stateMutability": "view",
-      "type": "function"
-   },
-   {
-      "inputs": [],
-      "name": "usdtToken",
-      "outputs": [
-         {
-            "internalType": "contract IERC20",
-            "name": "",
-            "type": "address"
-         }
-      ],
-      "stateMutability": "view",
-      "type": "function"
-   }
-]
