--- packages/dapp/src/utils/prod/abis/EventOrganizerService.json
@@ -1,310 +1,0 @@
-[
-   {
-      "inputs": [
-         {
-            "internalType": "contract Museum",
-            "name": "_museum",
-            "type": "address"
-         },
-         {
-            "internalType": "contract IERC20",
-            "name": "_usdtToken",
-            "type": "address"
-         }
-      ],
-      "stateMutability": "nonpayable",
-      "type": "constructor"
-   },
-   {
-      "inputs": [
-         {
-            "internalType": "address",
-            "name": "owner",
-            "type": "address"
-         }
-      ],
-      "name": "OwnableInvalidOwner",
-      "type": "error"
-   },
-   {
-      "inputs": [
-         {
-            "internalType": "address",
-            "name": "account",
-            "type": "address"
-         }
-      ],
-      "name": "OwnableUnauthorizedAccount",
-      "type": "error"
-   },
-   {
-      "anonymous": false,
-      "inputs": [
-         {
-            "indexed": true,
-            "internalType": "address",
-            "name": "artifactNFTAddress",
-            "type": "address"
-         },
-         {
-            "indexed": false,
-            "internalType": "string",
-            "name": "name",
-            "type": "string"
-         },
-         {
-            "indexed": false,
-            "internalType": "string",
-            "name": "symbol",
-            "type": "string"
-         },
-         {
-            "indexed": true,
-            "internalType": "address",
-            "name": "ownerAddress",
-            "type": "address"
-         },
-         {
-            "indexed": false,
-            "internalType": "string",
-            "name": "baseURI",
-            "type": "string"
-         }
-      ],
-      "name": "ArtifactNFTDeployed",
-      "type": "event"
-   },
-   {
-      "anonymous": false,
-      "inputs": [
-         {
-            "indexed": false,
-            "internalType": "string",
-            "name": "exhibitId",
-            "type": "string"
-         },
-         {
-            "indexed": true,
-            "internalType": "address",
-            "name": "exhibitNFTAddress",
-            "type": "address"
-         },
-         {
-            "indexed": true,
-            "internalType": "address",
-            "name": "escrowAddress",
-            "type": "address"
-         },
-         {
-            "indexed": true,
-            "internalType": "address",
-            "name": "museumAddress",
-            "type": "address"
-         }
-      ],
-      "name": "ExhibitNFTDeployed",
-      "type": "event"
-   },
-   {
-      "anonymous": false,
-      "inputs": [
-         {
-            "indexed": true,
-            "internalType": "address",
-            "name": "previousOwner",
-            "type": "address"
-         },
-         {
-            "indexed": true,
-            "internalType": "address",
-            "name": "newOwner",
-            "type": "address"
-         }
-      ],
-      "name": "OwnershipTransferred",
-      "type": "event"
-   },
-   {
-      "inputs": [
-         {
-            "internalType": "string",
-            "name": "name",
-            "type": "string"
-         },
-         {
-            "internalType": "string",
-            "name": "symbol",
-            "type": "string"
-         },
-         {
-            "internalType": "address",
-            "name": "owner",
-            "type": "address"
-         },
-         {
-            "internalType": "string",
-            "name": "baseURI",
-            "type": "string"
-         }
-      ],
-      "name": "deployArtifactNFT",
-      "outputs": [],
-      "stateMutability": "nonpayable",
-      "type": "function"
-   },
-   {
-      "inputs": [
-         {
-            "internalType": "string",
-            "name": "",
-            "type": "string"
-         }
-      ],
-      "name": "exhibits",
-      "outputs": [
-         {
-            "internalType": "address",
-            "name": "",
-            "type": "address"
-         }
-      ],
-      "stateMutability": "view",
-      "type": "function"
-   },
-   {
-      "inputs": [
-         {
-            "internalType": "string",
-            "name": "_id",
-            "type": "string"
-         }
-      ],
-      "name": "getExhibitNFTAddress",
-      "outputs": [
-         {
-            "internalType": "address",
-            "name": "",
-            "type": "address"
-         }
-      ],
-      "stateMutability": "view",
-      "type": "function"
-   },
-   {
-      "inputs": [],
-      "name": "museum",
-      "outputs": [
-         {
-            "internalType": "contract Museum",
-            "name": "",
-            "type": "address"
-         }
-      ],
-      "stateMutability": "view",
-      "type": "function"
-   },
-   {
-      "inputs": [
-         {
-            "internalType": "string",
-            "name": "name",
-            "type": "string"
-         },
-         {
-            "internalType": "string",
-            "name": "symbol",
-            "type": "string"
-         },
-         {
-            "internalType": "uint256",
-            "name": "ticketPrice",
-            "type": "uint256"
-         },
-         {
-            "internalType": "address[]",
-            "name": "beneficiaries",
-            "type": "address[]"
-         },
-         {
-            "internalType": "uint256[]",
-            "name": "shares",
-            "type": "uint256[]"
-         },
-         {
-            "internalType": "string",
-            "name": "baseURI",
-            "type": "string"
-         },
-         {
-            "internalType": "string",
-            "name": "location",
-            "type": "string"
-         },
-         {
-            "internalType": "address",
-            "name": "artifactNFTAddress",
-            "type": "address"
-         },
-         {
-            "internalType": "string",
-            "name": "details",
-            "type": "string"
-         },
-         {
-            "internalType": "string",
-            "name": "exhibitId",
-            "type": "string"
-         }
-      ],
-      "name": "organizeExhibit",
-      "outputs": [],
-      "stateMutability": "nonpayable",
-      "type": "function"
-   },
-   {
-      "inputs": [],
-      "name": "owner",
-      "outputs": [
-         {
-            "internalType": "address",
-            "name": "",
-            "type": "address"
-         }
-      ],
-      "stateMutability": "view",
-      "type": "function"
-   },
-   {
-      "inputs": [],
-      "name": "renounceOwnership",
-      "outputs": [],
-      "stateMutability": "nonpayable",
-      "type": "function"
-   },
-   {
-      "inputs": [
-         {
-            "internalType": "address",
-            "name": "newOwner",
-            "type": "address"
-         }
-      ],
-      "name": "transferOwnership",
-      "outputs": [],
-      "stateMutability": "nonpayable",
-      "type": "function"
-   },
-   {
-      "inputs": [],
-      "name": "usdtToken",
-      "outputs": [
-         {
-            "internalType": "contract IERC20",
-            "name": "",
-            "type": "address"
-         }
-      ],
-      "stateMutability": "view",
-      "type": "function"
-   }
-]
