--- packages/dapp/src/utils/prod/abis/Museum.json
@@ -1,254 +1,0 @@
-[
-   {
-      "inputs": [
-         {
-            "internalType": "contract IERC20",
-            "name": "_usdtToken",
-            "type": "address"
-         }
-      ],
-      "stateMutability": "nonpayable",
-      "type": "constructor"
-   },
-   {
-      "inputs": [
-         {
-            "internalType": "address",
-            "name": "owner",
-            "type": "address"
-         }
-      ],
-      "name": "OwnableInvalidOwner",
-      "type": "error"
-   },
-   {
-      "inputs": [
-         {
-            "internalType": "address",
-            "name": "account",
-            "type": "address"
-         }
-      ],
-      "name": "OwnableUnauthorizedAccount",
-      "type": "error"
-   },
-   {
-      "anonymous": false,
-      "inputs": [
-         {
-            "indexed": false,
-            "internalType": "address",
-            "name": "museumAddress",
-            "type": "address"
-         },
-         {
-            "indexed": false,
-            "internalType": "string",
-            "name": "exhibitId",
-            "type": "string"
-         },
-         {
-            "indexed": false,
-            "internalType": "address",
-            "name": "exhibitAddress",
-            "type": "address"
-         }
-      ],
-      "name": "ExhibitCurated",
-      "type": "event"
-   },
-   {
-      "anonymous": false,
-      "inputs": [
-         {
-            "indexed": false,
-            "internalType": "address",
-            "name": "museumAddress",
-            "type": "address"
-         },
-         {
-            "indexed": false,
-            "internalType": "address",
-            "name": "tokenAddress",
-            "type": "address"
-         },
-         {
-            "indexed": false,
-            "internalType": "address",
-            "name": "ownerAddress",
-            "type": "address"
-         }
-      ],
-      "name": "MuseumCreated",
-      "type": "event"
-   },
-   {
-      "anonymous": false,
-      "inputs": [
-         {
-            "indexed": true,
-            "internalType": "address",
-            "name": "previousOwner",
-            "type": "address"
-         },
-         {
-            "indexed": true,
-            "internalType": "address",
-            "name": "newOwner",
-            "type": "address"
-         }
-      ],
-      "name": "OwnershipTransferred",
-      "type": "event"
-   },
-   {
-      "anonymous": false,
-      "inputs": [
-         {
-            "indexed": false,
-            "internalType": "address",
-            "name": "buyer",
-            "type": "address"
-         },
-         {
-            "indexed": false,
-            "internalType": "address",
-            "name": "exhibit",
-            "type": "address"
-         },
-         {
-            "indexed": false,
-            "internalType": "uint256",
-            "name": "tokenId",
-            "type": "uint256"
-         }
-      ],
-      "name": "TicketPurchased",
-      "type": "event"
-   },
-   {
-      "inputs": [
-         {
-            "internalType": "string",
-            "name": "exhibitId",
-            "type": "string"
-         },
-         {
-            "internalType": "contract ExhibitNFT",
-            "name": "exhibit",
-            "type": "address"
-         }
-      ],
-      "name": "curateExhibit",
-      "outputs": [],
-      "stateMutability": "nonpayable",
-      "type": "function"
-   },
-   {
-      "inputs": [
-         {
-            "internalType": "string",
-            "name": "",
-            "type": "string"
-         }
-      ],
-      "name": "exhibits",
-      "outputs": [
-         {
-            "internalType": "contract ExhibitNFT",
-            "name": "",
-            "type": "address"
-         }
-      ],
-      "stateMutability": "view",
-      "type": "function"
-   },
-   {
-      "inputs": [],
-      "name": "owner",
-      "outputs": [
-         {
-            "internalType": "address",
-            "name": "",
-            "type": "address"
-         }
-      ],
-      "stateMutability": "view",
-      "type": "function"
-   },
-   {
-      "inputs": [
-         {
-            "internalType": "string",
-            "name": "exhibitId",
-            "type": "string"
-         },
-         {
-            "internalType": "uint256",
-            "name": "usdtAmount",
-            "type": "uint256"
-         }
-      ],
-      "name": "purchaseTicket",
-      "outputs": [],
-      "stateMutability": "nonpayable",
-      "type": "function"
-   },
-   {
-      "inputs": [],
-      "name": "renounceOwnership",
-      "outputs": [],
-      "stateMutability": "nonpayable",
-      "type": "function"
-   },
-   {
-      "inputs": [
-         {
-            "internalType": "address",
-            "name": "newOwner",
-            "type": "address"
-         }
-      ],
-      "name": "transferOwnership",
-      "outputs": [],
-      "stateMutability": "nonpayable",
-      "type": "function"
-   },
-   {
-      "inputs": [],
-      "name": "usdtToken",
-      "outputs": [
-         {
-            "internalType": "contract IERC20",
-            "name": "",
-            "type": "address"
-         }
-      ],
-      "stateMutability": "view",
-      "type": "function"
-   },
-   {
-      "inputs": [
-         {
-            "internalType": "string",
-            "name": "exhibitId",
-            "type": "string"
-         },
-         {
-            "internalType": "address",
-            "name": "user",
-            "type": "address"
-         }
-      ],
-      "name": "verifyTicketOwnership",
-      "outputs": [
-         {
-            "internalType": "bool",
-            "name": "",
-            "type": "bool"
-         }
-      ],
-      "stateMutability": "view",
-      "type": "function"
-   }
-]
