--- packages/dapp/src/utils/prod/uris/0
@@ -1,66 +1,0 @@
-{
-    "id": "1",
-    "name": "1912.06.0492 :: Double Drum with Zigzag Pattern",
-    "image": "https://s3.tebi.io/summitshare-previews/1.png",
-    "animation_url": "https://s3.tebi.io/summitshare-artifacts/1.glb",
-    "attributes": [
-        {
-            "trait_type": "Category",
-            "value": "Musical Instrument"
-        },
-        {
-            "trait_type": "Material",
-            "value": "Wood, Varanus Skin"
-        },
-        {
-            "trait_type": "Ethnic Group",
-            "value": "Bisa"
-        },
-        {
-            "trait_type": "Country of Origin",
-            "value": "Zambia"
-        },
-        {
-            "trait_type": "Description",
-            "value": "Double drum with a zigzag pattern in the middle, Varanus skin on both ends."
-        },
-        {
-            "trait_type": "Dimensions",
-            "value": "Length 65 cm, Diameter 31 cm"
-        },
-        {
-            "trait_type": "Collector",
-            "value": "Eric von Rosen"
-        },
-        {
-            "trait_type": "Acquisition Year",
-            "value": "1912"
-        },
-        {
-            "trait_type": "Expedition",
-            "value": "Swedish Rhodesia-Congo Expedition (1911-1912)"
-        },
-        {
-            "trait_type": "Special Features",
-            "value": "Incorporates sand within to produce a hissing sound when tilted."
-        }
-    ],
-    "description": "This double drum, collected by Eric von Rosen during the Swedish Rhodesia-Congo Expedition of 1911-1912, showcases a unique zigzag pattern with Varanus skin at both ends. The drum is a significant artifact representing the Bisa people of Zambia, highlighting their musical heritage. Von Rosen, a renowned explorer and ethnographer, contributed significantly to the understanding of Central African cultures through his meticulous documentation and collection of artifacts, such as this double drum. The drum not only serves as a musical instrument but also as a piece of cultural history, illustrating the craftsmanship and artistic expression of the Bisa people.",
-    "external_url": "https://collections.smvk.se/carlotta-em/web/object/1996946",
-    "collector_info": {
-        "name": "Eric von Rosen",
-        "lifetime": "1879 - 1948",
-        "roles": ["Explorer", "Ethnographer"],
-        "background": "Von Rosen's expeditions across Africa, especially in Zambia, were pivotal in bringing to light the diverse cultures and artifacts of the continent. His work has been instrumental in preserving African heritage and providing a window into the lives of people like the Bisa."
-    },
-    "publication_info": {
-        "reference": "Rosen, Eric von (1916). Träskfolket: svenska Rhodesia-Kongo-expeditionens etnografiska forskningsresultat.",
-        "description": "This publication by Eric von Rosen provides in-depth insights into the ethnographic findings from the Swedish Rhodesia-Congo Expedition, including detailed descriptions of artifacts such as this double drum."
-    },
-    "event_info": {
-        "event": "Rhodesia-Congo Expedition Exhibition",
-        "start_date": "1913-02-04",
-        "description": "This exhibition showcased the findings from the Swedish Rhodesia-Congo Expedition, offering a comprehensive view of the artifacts collected, including the double drum, and highlighting the rich cultural heritage of the African continent."
-    }
-}
-
