--- packages/dapp/src/utils/prod/uris/1
@@ -1,65 +1,0 @@
-{
-    "id": "2",
-    "name": "1894.15.0035 :: Girdle/Belt with Cowrie Shells",
-    "image": "https://s3.tebi.io/summitshare-previews/2.png",
-    "animation_url": "https://s3.tebi.io/summitshare-artifacts/2.glb",
-    "attributes": [
-        {
-            "trait_type": "Category",
-            "value": "Girdle/Belt"
-        },
-        {
-            "trait_type": "Material",
-            "value": "Leather; Cowrie Shells"
-        },
-        {
-            "trait_type": "Ethnic Group",
-            "value": "Mbunda"
-        },
-        {
-            "trait_type": "Country of Origin",
-            "value": "Zambia"
-        },
-        {
-            "trait_type": "Description",
-            "value": "Short leather strap girdle/belt adorned with sewn-on cowrie shells."
-        },
-        {
-            "trait_type": "Length",
-            "value": "49.5 cm"
-        },
-        {
-            "trait_type": "Width",
-            "value": "0.5 cm"
-        },
-        {
-            "trait_type": "Collector",
-            "value": "Emil Holub"
-        },
-        {
-            "trait_type": "Acquisition Year",
-            "value": "1894"
-        }
-    ],
-    "description": "This girdle or belt, adorned with cowrie shells and dating back to 1894, is a testament to the craftsmanship and cultural heritage of the Mbunda people from Zambia. Collected by Emil Holub, a Czech explorer and ethnographer, during his extensive travels in Africa, this item represents the rich cultural exchanges between Africa and Europe in the late 19th century. Holub's contributions to ethnographic collections have provided invaluable insights into the material culture of Central and Southern African societies.",
-    "external_url": "https://collections.smvk.se/carlotta-em/web/object/1600518",
-    "collector_info": {
-        "name": "Emil Holub",
-        "lifetime": "1847 - 1902",
-        "roles": ["Explorer", "Ethnographer"],
-        "background": "Known for his travels across Africa, Holub contributed significantly to the understanding of African cultures. His work, including a notable expedition to the Zambezi and Congo rivers' sources, has left a lasting legacy in ethnographic studies."
-    },
-    "event_info": [
-        {
-            "event": "Collection visit by Mulenga Kapwepwe and Samba Yonga",
-            "date": "2020-09-18",
-            "description": "A visit aimed at studying Zambian artifacts within the framework of 'Empowering Women's Histories: Sharing Digital Heritage between Zambia and Sweden', highlighting the collaborative effort to make collections more accessible and relevant through digital means."
-        },
-        {
-            "event": "Collection visit by Victoria Phiri Chitungu",
-            "date": "2022-11-08",
-            "description": "Part of the ongoing project 'Empowering Women's Histories', this visit focused on the examination and documentation of Zambian everyday objects, underlining the importance of preserving and sharing cultural heritage."
-        }
-    ]
-}
-
