--- packages/dapp/src/utils/prod/uris/2
@@ -1,50 +1,0 @@
-{
-    "id": "3",
-    "name": "1909.17.0012 :: Snuff Box",
-    "image": "https://s3.tebi.io/summitshare-previews/3.png",
-    "animation_url": "https://s3.tebi.io/summitshare-artifacts/3.glb",
-    "attributes": [
-        {
-            "trait_type": "Category",
-            "value": "Snuff Box"
-        },
-        {
-            "trait_type": "Country of Origin",
-            "value": "Zambia"
-        },
-        {
-            "trait_type": "Material",
-            "value": "Mixed Material"
-        },
-        {
-            "trait_type": "Collector",
-            "value": "Magnus Leijer"
-        },
-        {
-            "trait_type": "Acquisition Year",
-            "value": "1909"
-        },
-        {
-            "trait_type": "Description",
-            "value": "Snuff box according to the general catalog."
-        },
-        {
-            "trait_type": "Special Features",
-            "value": "Used for holding snuff, showcasing craftsmanship from Zambia."
-        }
-    ],
-    "description": "This snuff box, collected by Magnus Leijer, reflects the intricate craftsmanship of Zambia. Leijer, a prominent figure known for his contributions as a big game hunter, vice consul, and a significant collector of African artifacts, acquired this piece in 1909. The box symbolizes the cultural heritage and practices surrounding snuff use in Zambia. Furthermore, the piece was part of a collection visit by Victoria Phiri Chitungu in 2022, underlining the ongoing exploration and sharing of Zambian and Swedish heritage through projects like 'Empowering Women's Histories: Sharing Digital Heritage between Zambia and Sweden'.",
-    "external_url": "https://collections.smvk.se/carlotta-em/web/object/1304494",
-    "collector_info": {
-        "name": "Magnus Leijer",
-        "lifetime": "1869 - 1946",
-        "roles": ["Big Game Hunter", "Vice Consul", "Collector"],
-        "background": "Leijer's career spanned roles as a factory manager in Cameroon, a civil administrator in Northern Rhodesia (now Zambia), and a British vice-consul. His contributions to collections of African artifacts have been significant, particularly those acquired during his time in Africa, which include not only this snuff box but also various everyday objects reflecting the social history and material culture of the region."
-    },
-    "event_info": {
-        "event": "Collection visit by Victoria Phiri Chitungu",
-        "date": "2022-11-08",
-        "description": "Part of the 'Empowering Women's Histories: Sharing Digital Heritage between Zambia and Sweden' project, this visit involved the examination of physical objects to deepen understanding of their manufacture, material, ornamentation, and both historical and contemporary use."
-    }
-}
-
