--- packages/dapp/src/utils/prod/uris/3
@@ -1,57 +1,0 @@
-{
-    "id": "4",
-    "name": "1912.06.0132 :: Headrest",
-    "image": "https://s3.tebi.io/summitshare-previews/4.png",
-    "animation_url": "https://s3.tebi.io/summitshare-artifacts/4.glb",
-    "attributes": [
-        {
-            "trait_type": "Category",
-            "value": "Headrest"
-        },
-        {
-            "trait_type": "Ethnic Group",
-            "value": "Twa"
-        },
-        {
-            "trait_type": "Region of Origin",
-            "value": "Lake Bangweulu"
-        },
-        {
-            "trait_type": "Country of Origin",
-            "value": "Zambia"
-        },
-        {
-            "trait_type": "Material",
-            "value": "Wood"
-        },
-        {
-            "trait_type": "Collector",
-            "value": "Eric von Rosen"
-        },
-        {
-            "trait_type": "Acquisition Year",
-            "value": "1912"
-        },
-        {
-            "trait_type": "Description",
-            "value": "Wooden stool, with carved ornaments."
-        },
-        {
-            "trait_type": "Expedition",
-            "value": "Swedish Rhodesia-Congo Expedition (1911-1912)"
-        },
-        {
-            "trait_type": "Additional Info",
-            "value": "Features carved ornaments often used by the swamp people, including a net-like pattern formed by diamond cuts and parallel zigzag lines."
-        }
-    ],
-    "description": "This headrest, collected by Eric von Rosen during the Swedish Rhodesia-Congo Expedition of 1911-1912, is crafted from wood with intricate carved ornaments. Originating from the Twa people near Lake Bangweulu in Zambia, it showcases the detailed craftsmanship and cultural heritage of the region. Von Rosen's expedition contributed significantly to the documentation and preservation of Central African cultures, bringing back a wealth of ethnographic items to Sweden.",
-    "external_url": "https://collections.smvk.se/carlotta-em/web/object/1261202",
-    "collector_info": {
-        "name": "Eric von Rosen",
-        "lifetime": "1879 - 1948",
-        "roles": ["Explorer", "Ethnographer"],
-        "contributions": "Documented the Twa people and their artifacts during the Swedish Rhodesia-Congo Expedition, enriching the Swedish National Museum of World Culture with a significant collection of ethnographic items."
-    }
-}
-
