--- packages/dapp/src/utils/prod/uris/4
@@ -1,54 +1,0 @@
-{
-    "id": "5",
-    "name": "1912.06.0417 :: Calabash - Gourd Water Bottle",
-    "image": "https://s3.tebi.io/summitshare-previews/5.png",
-    "animation_url": "https://s3.tebi.io/summitshare-artifacts/5.glb",
-    "attributes": [
-        {
-            "trait_type": "Category",
-            "value": "Vessel"
-        },
-        {
-            "trait_type": "Material",
-            "value": "Gourd, Bast"
-        },
-        {
-            "trait_type": "Ethnic Group",
-            "value": "Bisa"
-        },
-        {
-            "trait_type": "Country of Origin",
-            "value": "Zambia"
-        },
-        {
-            "trait_type": "Description",
-            "value": "Large water bottle of gourd with burnt ornaments, wrapped with bast for easier carrying."
-        },
-        {
-            "trait_type": "Collector",
-            "value": "Eric von Rosen"
-        },
-        {
-            "trait_type": "Acquisition Year",
-            "value": "1912"
-        },
-        {
-            "trait_type": "Expedition",
-            "value": "Swedish Rhodesia-Congo Expedition (1911-1912)"
-        }
-    ],
-    "description": "This large gourd water bottle, collected by Eric von Rosen during the Swedish Rhodesia-Congo Expedition of 1911-1912, is a remarkable artifact from the Bisa people of Zambia. The vessel's surface features burnt ornaments, a traditional decorative technique, and is wrapped with bast to facilitate carrying. This bottle exemplifies the ingenuity and artistry of the Bisa people in creating functional yet aesthetically pleasing objects for daily use.",
-    "external_url": "https://collections.smvk.se/carlotta-em/web/object/1996935",
-    "collector_info": {
-        "name": "Eric von Rosen",
-        "lifetime": "1879 - 1948",
-        "roles": ["Explorer", "Ethnographer"],
-        "background": "Eric von Rosen was a notable figure in early 20th-century exploration and ethnography. His contributions to the documentation of African cultures, including the Bisa people's way of life, have provided valuable insights into the diverse cultural heritage of Zambia. The collection and donation of artifacts like this gourd water bottle to museums have allowed for the preservation and study of African material culture."
-    },
-    "event_info": {
-        "event": "Rhodesia-Congo Expedition Exhibition",
-        "start_date": "1913-02-04",
-        "description": "The exhibition showcased the findings from the Swedish Rhodesia-Congo Expedition, highlighting the diverse cultures and artifacts collected by Eric von Rosen and his team. This gourd water bottle was among the many items presented, offering a glimpse into the daily life and craftsmanship of the Bisa people."
-    }
-}
-
