--- packages/dapp/src/utils/prod/uris/5
@@ -1,58 +1,0 @@
-{
-    "id": "6",
-    "name": "1969.29.0062 :: Luwale Mask",
-    "image": "https://s3.tebi.io/summitshare-previews/6.png",
-    "animation_url": "https://s3.tebi.io/summitshare-artifacts/6.glb",
-    "attributes": [
-        {
-            "trait_type": "Category",
-            "value": "Mask"
-        },
-        {
-            "trait_type": "Material",
-            "value": "Wood, Bark, Strings"
-        },
-        {
-            "trait_type": "Ethnic Group",
-            "value": "Luvale"
-        },
-        {
-            "trait_type": "Country of Origin",
-            "value": "Zambia"
-        },
-        {
-            "trait_type": "Description",
-            "value": "Wooden mask with red-painted face, dot tattoo, half-open mouth with teeth marks, and blue-black strings for hair."
-        },
-        {
-            "trait_type": "Dimensions",
-            "value": "Height: 20 cm, Width: 15 cm, Hair Length: 55 cm"
-        },
-        {
-            "trait_type": "Collector",
-            "value": "John-Erik Elmberg"
-        },
-        {
-            "trait_type": "Acquisition Year",
-            "value": "1969"
-        },
-        {
-            "trait_type": "Fieldwork",
-            "value": "Fieldwork by John-Erik Elmberg in Zambia, 1969"
-        }
-    ],
-    "description": "This Luwale mask, collected by John-Erik Elmberg during his 1969 fieldwork in Zambia, is a profound representation of the Luvale people's artistic and cultural expressions. Elmberg, a significant figure in ethnographic studies, conducted extensive research across various regions, including Africa. His work has contributed valuable insights into the cultures of the Zambian people. The mask, characterized by its intricate design and symbolic decorations, was part of Elmberg's collections donated to the Ethnographic Museum, showcasing the diverse heritage of Zambia.",
-    "external_url": "https://collections.smvk.se/carlotta-em/web/object/2042752",
-    "collector_info": {
-        "name": "John-Erik Elmberg",
-        "lifetime": "1917 - 1970",
-        "roles": ["Docent", "Researcher"],
-        "background": "Elmberg's scholarly work and field research in Zambia and other locations have greatly contributed to the understanding of social changes and agricultural development in post-colonial Africa. His collections offer a glimpse into the daily life and spiritual practices of various ethnic groups."
-    },
-    "event_info": {
-        "event": "Collection visit by the Women's History Museum Zambia",
-        "date": "2020-09-18",
-        "description": "The visit by Mulenga Kapwepwe and Samba Yonga aimed to explore Zambian artifacts within the 'Empowering women's histories: sharing digital heritage between Zambia and Sweden' project. This collaboration seeks to make cultural heritage more accessible and relevant through digital tools, emphasizing women's histories."
-    }
-}
-
